pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transport-shipment-etl"
version = "0.1.0"
description = "Transportation shipment ETL pipeline using PySpark, Spark SQL, curated Hive-style outputs, and EMR deployment artifacts"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Transportation Data Engineering Team" }]
dependencies = [
  "pyspark==3.5.2",
  "PyYAML==6.0.2",
  "python-dateutil==2.9.0.post0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==6.0.0",
  "black==24.10.0",
  "ruff==0.11.0",
  "mypy==1.15.0",
]

[project.scripts]
transport-etl = "transport_etl.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.black]
line-length = 100
